Modified the Design Branch to work on Old Phone.
- This branch was synced to the latest version of the master branch, and
was modified so that it works on the older phone. The idea being that I
can modify the code on the master branch, and then merge it with this
one to test it on the phone, and hopefully not have to mess around with
commenting out code every time I want to test it.